Optimize for mobile
The previous build had problems on mobile as the kenji-bedtime image was
too wide and caused a horizontal scrollbar to appear. I added a media
query to resolve this on thinner screens.
Mobile viewing also made the button in the call to action section quite
thin and tall due to the flexbox pushing it. I resolved this issue using
flex-wrap.
This commit optimizes index.html for viewing on mobile with these two
changes.